Move API validations package
