+++ pkg/apis/certmanager/BUILD.bazel
@@ -20,7 +20,6 @@ filegroup(
         ":package-srcs",
         "//pkg/apis/certmanager/install:all-srcs",
         "//pkg/apis/certmanager/v1alpha1:all-srcs",
-        "//pkg/apis/certmanager/validation:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
