+++ pkg/controller/certificaterequests/BUILD.bazel
@@ -13,11 +13,11 @@ go_library(
         "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager:go_default_library",
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
-        "//pkg/apis/certmanager/validation:go_default_library",
         "//pkg/client/clientset/versioned:go_default_library",
         "//pkg/client/listers/certmanager/v1alpha1:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/controller/certificaterequests/util:go_default_library",
+        "//pkg/internal/apis/certmanager/validation:go_default_library",
         "//pkg/issuer:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/metrics:go_default_library",
