+++ pkg/internal/apis/certmanager/validation/util/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["nameserver.go"],
-    importpath = "github.com/jetstack/cert-manager/pkg/apis/certmanager/validation/util",
+    importpath = "github.com/jetstack/cert-manager/pkg/internal/apis/certmanager/validation/util",
     visibility = ["//visibility:public"],
 )
