+++ pkg/webhook/BUILD.bazel
@@ -10,8 +10,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
-        "//pkg/apis/certmanager/validation:go_default_library",
         "//pkg/internal/apis/certmanager/install:go_default_library",
+        "//pkg/internal/apis/certmanager/validation:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
     ],
 )
